refactor: update import paths for UserOption, UserAvatar, and userOptions in UserPage component
