+++ src/user/interfaces/UserOption.ts
@@ -1,0 +1,9 @@
+import { IconProp } from "@fortawesome/fontawesome-svg-core";
+
+export default interface UserOption {
+  href: string;
+  icon: IconProp;
+  iconColor: string;
+  title: string;
+  description: string;
+}
